# Conflicts:
#	main.py
#	models/domain/dessert.py
#	models/domain/drink.py
#	models/domain/ingredient.py
#	models/domain/order.py
#	models/pizza.py
#	modules/mysql_model.py

+++ models/domain/drink.py
@@ -1,7 +1,7 @@
-from models.mysql_model import db
+from modules.mysql_model import db
 
 
 class Drink(db.Model):
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     name = db.Column(db.String(80), unique=True, nullable=False)
-    price = db.Column(db.Integer, nullable=False)
+    price = db.Column(db.Float, nullable=False)
